PTG: new isBijective() method
Plus, more ASSERTS, etc.

+++ libs/nav/include/mrpt/nav/tpspace/CPTG_Holo_Blend.h
@@ -51,8 +51,8 @@ namespace mrpt
 		virtual mrpt::kinematics::CVehicleVelCmdPtr getSupportedKinematicVelocityCommand() const MRPT_OVERRIDE;
 
 		size_t getPathStepCount(uint16_t k) const MRPT_OVERRIDE;
-		void getPathPose(uint16_t k, uint16_t step, mrpt::math::TPose2D &p) const MRPT_OVERRIDE;
-		double getPathDist(uint16_t k, uint16_t step) const  MRPT_OVERRIDE;
+		void getPathPose(uint16_t k, uint32_t step, mrpt::math::TPose2D &p) const MRPT_OVERRIDE;
+		double getPathDist(uint16_t k, uint32_t step) const  MRPT_OVERRIDE;
 		bool getPathStepForDist(uint16_t k, double dist, uint32_t &out_step) const MRPT_OVERRIDE;
 		double getPathStepDuration() const MRPT_OVERRIDE;
 		double getMaxLinVel() const MRPT_OVERRIDE { return V_MAX; }
